rename class to follow the project standard
